+++ src/detections/detection.rs
@@ -191,6 +191,7 @@ impl Detection {
                 record_info,
                 stored_static.verbose_flag,
                 stored_static.quiet_errors_flag,
+                stored_static.json_input_flag,
                 &stored_static.eventkey_alias,
             );
             if !result {
@@ -240,7 +241,8 @@ impl Detection {
         );
 
         let default_time = Utc.with_ymd_and_hms(1970, 1, 1, 0, 0, 0).unwrap();
-        let time = message::get_event_time(&record_info.record).unwrap_or(default_time);
+        let time = message::get_event_time(&record_info.record, stored_static.json_input_flag)
+            .unwrap_or(default_time);
         let level = rule.yaml["level"].as_str().unwrap_or("-");
 
         let mut profile_converter: HashMap<&str, Profile> = HashMap::new();
@@ -274,7 +276,7 @@ impl Detection {
                             stored_static
                                 .ch_config
                                 .get(&ch_str.to_ascii_lowercase())
-                                .unwrap_or(ch_str),
+                                .unwrap_or(&ch_str.to_string()),
                         )),
                     );
                 }
@@ -842,6 +844,8 @@ impl Detection {
             format!("Total enabled detection rules: {}", total_loaded_rule_cnt);
         println!("{}", tmp_total_detect_output);
         println!();
+        println!("Scanning in progress. Please wait.");
+        println!();
         if stored_static.html_report_flag {
             html_report_stock.push(format!("- {}", tmp_total_detect_output));
         }
@@ -903,6 +907,7 @@ mod tests {
                     html_report: None,
                     no_summary: false,
                 },
+                json_input: false,
             })),
             no_color: false,
             quiet: false,