+++ src/detections/rule/matchers.rs
@@ -654,6 +654,7 @@ mod tests {
                     html_report: None,
                     no_summary: false,
                 },
+                json_input: false,
             })),
             no_color: false,
             quiet: false,
@@ -671,6 +672,7 @@ mod tests {
                         &recinfo,
                         dummy_stored_static.verbose_flag,
                         dummy_stored_static.quiet_errors_flag,
+                        dummy_stored_static.json_input_flag,
                         &dummy_stored_static.eventkey_alias
                     ),
                     expect_select
@@ -1675,7 +1677,7 @@ mod tests {
         enabled: true
         detection:
             selection:
-                Channel: 
+                Channel:
                     value: Security
         details: 'command=%CommandLine%'
         "#;
@@ -1697,7 +1699,7 @@ mod tests {
         enabled: true
         detection:
             selection:
-                Channel: 
+                Channel:
                     value: Securiteen
         details: 'command=%CommandLine%'
         "#;
@@ -1904,9 +1906,9 @@ mod tests {
         enabled: true
         detection:
             selection:
-                Channel: 
+                Channel:
                     value: Security
-                Takoyaki: 
+                Takoyaki:
                     value: null
         details: 'command=%CommandLine%'
         "#;