+++ src/detections/rule/selectionnodes.rs
@@ -281,7 +281,7 @@ impl LeafSelectionNode {
 
     /// JSON形式のEventJSONから値を取得する関数 aliasも考慮されている。
     fn get_event_value<'a>(&self, record: &'a EvtxRecordInfo) -> Option<&'a String> {
-        // keyが指定されたいない場合は
+        // keyが指定されていない場合はそのままのレコードのデータを取得する
         if self.key_list.is_empty() {
             return Option::Some(&record.data_string);
         }
@@ -477,6 +477,7 @@ mod tests {
                     html_report: None,
                     no_summary: false,
                 },
+                json_input: false,
             })),
             no_color: false,
             quiet: false,
@@ -498,6 +499,7 @@ mod tests {
                         &recinfo,
                         dummy_stored_static.verbose_flag,
                         dummy_stored_static.quiet_errors_flag,
+                        dummy_stored_static.json_input_flag,
                         &dummy_stored_static.eventkey_alias
                     ),
                     expect_select
@@ -559,7 +561,7 @@ mod tests {
         enabled: true
         detection:
             selection:
-                Channel: 
+                Channel:
                     - PowerShell
                     - Security
         details: 'command=%CommandLine%'
@@ -581,7 +583,7 @@ mod tests {
         enabled: true
         detection:
             selection:
-                Channel: 
+                Channel:
                     - PowerShell
                     - Security
         details: 'command=%CommandLine%'
@@ -603,7 +605,7 @@ mod tests {
         enabled: true
         detection:
             selection:
-                Channel: 
+                Channel:
                     - PowerShell
                     - Security
         details: 'command=%CommandLine%'