+++ src/options/htmlreport.rs
@@ -242,6 +242,7 @@ mod tests {
                 html_report: Some(Path::new("./dummy").to_path_buf()),
                 no_summary: false,
             },
+            json_input: false,
         });
         let csv_html_flag_enable = create_dummy_stored_static(Some(enable_csv_action));
         assert!(htmlreport::check_html_flag(&csv_html_flag_enable.config));
@@ -279,6 +280,7 @@ mod tests {
                 html_report: None,
                 no_summary: false,
             },
+            json_input: false,
         });
         let csv_html_flag_disable = create_dummy_stored_static(Some(disable_csv_action));
         assert!(!htmlreport::check_html_flag(&csv_html_flag_disable.config));
@@ -320,6 +322,7 @@ mod tests {
                 no_summary: false,
             },
             jsonl_timeline: false,
+            json_input: false,
         });
         let json_html_flag_enable = create_dummy_stored_static(Some(enable_json_action));
         assert!(htmlreport::check_html_flag(&json_html_flag_enable.config));
@@ -358,6 +361,7 @@ mod tests {
                 no_summary: false,
             },
             jsonl_timeline: false,
+            json_input: false,
         });
         let json_html_flag_disable = create_dummy_stored_static(Some(disable_json_action));
         assert!(!htmlreport::check_html_flag(&json_html_flag_disable.config));