+++ src/options/profile.rs
@@ -397,6 +397,7 @@ mod tests {
                     html_report: None,
                     no_summary: false,
                 },
+                json_input: false,
             }));
         assert_eq!(
             Nested::<Vec<String>>::new(),
@@ -495,6 +496,7 @@ mod tests {
                     html_report: None,
                     no_summary: false,
                 },
+                json_input: false,
             }));
         assert_eq!(
             Some(expect),
@@ -542,6 +544,7 @@ mod tests {
                     html_report: None,
                     no_summary: false,
                 },
+                json_input: false,
             }));
 
         let expect: Vec<(CompactString, Profile)> = vec![
@@ -620,6 +623,7 @@ mod tests {
                     html_report: None,
                     no_summary: false,
                 },
+                json_input: false,
             }));
         //両方のファイルが存在しない場合
         assert_eq!(